pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tradelunch-blog-agents"
version = "1.0.0"
description = "Multi-agent system for automating blog post processing using LangGraph and Qwen 2.5 8B"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "TradeLunch", email = "tio.taek.lim@gmail.com"}
]
keywords = ["ai", "agents", "blog", "automation", "langgraph", "llm", "qwen"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # SQLAlchemy with async PostgreSQL
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "greenlet>=3.0.0",

    # LangChain & LangGraph
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-ollama>=0.1.0",
    "langgraph>=0.0.40",

    # LLM - Local
    "ollama>=0.1.0",

    # LLM - Cloud APIs
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "openai>=1.12.0",
    "anthropic>=0.18.0",

    # Rich CLI
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",

    # Markdown processing
    "python-frontmatter>=1.0.0",
    "markdown>=3.5.0",

    # AWS (optional, for real S3/RDS)
    "boto3>=1.34.0",
    "psycopg2-binary>=2.9.0",

    # Utilities
    "python-dateutil>=2.8.0",
    "pydantic>=2.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[project.urls]
Homepage = "https://github.com/tradelunch/blog-agents"
Repository = "https://github.com/tradelunch/blog-agents"
Issues = "https://github.com/tradelunch/blog-agents/issues"

[project.scripts]
blog-agent = "cli_multi_agent:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "config", "schema"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | tradelunch-agents-venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # imported but unused

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=agents",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
]
testpaths = ["__tests__"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["agents"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/tradelunch-agents-venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
